+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ HEAD
 
   - extract cmd, args and workload to Executable class
   - remove ``--remove-all`` argument in ``perun rm``
+  - add support for removing profiles from pending jobs through perun
 
 0.17.1 (2019-07-24)
 -------------------